+++ task_management/client/bundles/containers/templates/DashboardContainer.jsx
@@ -2,20 +2,20 @@ import React, { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import styled from "styled-components";
 import Grid from "@material-ui/core/Grid";
-import { SearchContent } from "./SearchContent";
-import { GrPaper } from "./GrPaper";
-import { PaperHeader } from "./PaperHeader";
-import { PaperSearchContent } from "./PaperSearchContent";
-import { PaperBody } from "./PaperBody";
-import { AddButton } from "./AddButton";
-import { DashboardTable } from "./DashboardTable";
-import { ProjectThead } from "./ProjectThead";
-import { ProjectTableTr } from "./ProjectTableTr";
-import { MemberTableTr } from "./MemberTableTr";
+import { SearchContent } from "../../components/molecules/SearchContent";
+import { GrPaper } from "../../components/molecules/GrPaper";
+import { PaperHeader } from "../../components/molecules/PaperHeader";
+import { PaperSearchContent } from "../../components/organisms/PaperSearchContent";
+import { PaperBody } from "../../components/atoms/PaperBody";
+import { AddButton } from "../../components/atoms/AddButton";
+import { DashboardTable } from "../../components/organisms/DashboardTable";
+import { ProjectThead } from "../../components/molecules/ProjectThead";
+import { ProjectTableTr } from "../../components/molecules/ProjectTableTr";
+import { MemberTableTr } from "../../components/molecules/MemberTableTr";
 import DescriptionIcon from "@material-ui/icons/Description";
 import PeopleIcon from "@material-ui/icons/People";
-import { AddProjectModal } from "./AddProjectModal";
-import { AddUserModal } from "./AddUserModal";
+import { AddProjectModal } from "../../components/organisms/AddProjectModal";
+import { AddUserModal } from "../../components/organisms/AddUserModal";
 import gql from "graphql-tag";
 import { useMutation } from "@apollo/react-hooks";
