+++ task_management/client/bundles/containers/templates/LoginForm.jsx
@@ -10,7 +10,7 @@ import Paper from "@material-ui/core/Paper";
 import Grid from "@material-ui/core/Grid";
 import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
 import Typography from "@material-ui/core/Typography";
-import { Form } from "../../Shared/components/forms/Form";
+import { Form } from "../../components/atoms/forms/Form";
 import { makeStyles } from "@material-ui/core/styles";
 
 const useStyles = makeStyles(theme => ({
