+++ task_management/client/bundles/containers/templates/SignUp.jsx
@@ -9,7 +9,7 @@ import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
 import Container from "@material-ui/core/Container";
-import { Form } from "../../Shared/components/forms/Form";
+import { Form } from "../../components/atoms/forms/Form";
 
 const useStyles = makeStyles(theme => ({
   "@global": {
